Modify the k factor in ELO calculation
Since Elo was originally designed for one-on-one matches, the score
adjustment per game is a fixed k. In the current implementation, this k
is divided among all members of a team, so each player only gains a
very small amount.

This commit changes k into a value that scales with the number of
players instead of being fixed, in order to better balance team sizes.